+++ src/extension.ts
@@ -47,6 +47,7 @@ export async function activate(context: vscode.ExtensionContext) {
     vscode.commands.registerCommand('voxpilot.selectAudioDevice', () => engine?.selectAudioDevice()),
     vscode.commands.registerCommand('voxpilot.inlineVoiceInput', () => engine?.inlineVoiceInput()),
     vscode.commands.registerCommand('voxpilot.transcriptHistory', () => engine?.showTranscriptHistory()),
+    vscode.commands.registerCommand('voxpilot.openHistoryPanel', () => engine?.openHistoryPanel()),
     vscode.commands.registerCommand('voxpilot.sendToChat', () => engine?.sendLastToChat()),
     vscode.commands.registerCommand('voxpilot.selectLanguage', () => engine?.selectLanguage()),
     vscode.commands.registerCommand('voxpilot.quickToggleLanguage', () => engine?.quickToggleLanguage()),